backend/requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-vertexai
google-cloud-aiplatform